requirements.txt
django
djangorestframework
pillow
psycopg2-binary
python-decouple
django-filter
drf-yasg
gunicorn
djangorestframework-simplejwt
django-rest-passwordreset
coreapi
redis
celery
django-cachalot
docker
docker-compose
coreapi
django-cors-headers
lxml
beautifulsoup4
requests
aiohttp